+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for CLI interface.
 - Docstrings in dist modules.
+- String input.
 
 ### Fixed
 - CLI will now output a better error message when input is not a number.